requirements.txt
amqp==2.5.0
appnope==0.1.0
asn1crypto==0.24.0
backcall==0.1.0
billiard==3.6.0.0
celery==4.3.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
decorator==4.4.0
Django==2.2.1
django-cors-headers==3.0.2
django-filter==2.1.0
django-redis==4.10.0
djangorestframework==3.9.4
idna==2.8
ipython==7.5.0
ipython-genutils==0.2.0
jedi==0.14.1
Jinja2==2.10.1
kombu==4.6.3
MarkupSafe==1.1.1
mysqlclient==1.4.2.post1
optionaldict==0.1.1
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.4.2
python-dateutil==2.8.0
pytz==2019.2
redis==3.3.5
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
traitlets==4.3.2
ujson==1.35
urllib3==1.25.3
vine==1.3.0
wcwidth==0.1.7
wechatpy==1.8.3
xmltodict==0.12.0
